--- server/querier/app/distributed-tracing/service/tracemap/go.mod
@@ -1,4 +1,0 @@
-module github.com/deepflowio/deepflow/server/querier/app/distributed-tracing/service/tracemap
-
-go 1.18
-
