+++ server/querier/app/distributed_tracing/router/tracemap.go
@@ -20,9 +20,9 @@ import (
 	"github.com/gin-gonic/gin"
 	"github.com/gin-gonic/gin/binding"
 
-	"github.com/deepflowio/deepflow/server/querier/app/distributed-tracing/common"
-	"github.com/deepflowio/deepflow/server/querier/app/distributed-tracing/model"
-	"github.com/deepflowio/deepflow/server/querier/app/distributed-tracing/service/tracemap"
+	"github.com/deepflowio/deepflow/server/querier/app/distributed_tracing/common"
+	"github.com/deepflowio/deepflow/server/querier/app/distributed_tracing/model"
+	"github.com/deepflowio/deepflow/server/querier/app/distributed_tracing/service/tracemap"
 	"github.com/deepflowio/deepflow/server/querier/config"
 	"github.com/deepflowio/deepflow/server/querier/router"
 )
