+++ server/querier/app/distributed_tracing/service/tracemap/go.mod
@@ -1,0 +1,4 @@
+module github.com/deepflowio/deepflow/server/querier/app/distributed_tracing/service/tracemap
+
+go 1.18
+
